# services/aisimulators/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AISimulators API
  description: |
    Minimal REST API for LLM inference GPU sizing and memory estimation.
    Wraps the aisimulate SDK (`cli_recommend` and `estimate_kv_cache`).
  version: 1.0.0
  contact:
    name: ConfigIQ team
servers:
  - url: http://localhost:7860
    description: Local development
  - url: https://aisimulators.dev
    description: Production
...
paths:
  /recommend:
    post:
      operationId: recommend
      tags: [sizing]
      summary: Find optimal GPU configuration for a workload
      description: |
        Given a model, GPU system, backend, workload shape, and performance
        targets (SLA), searches across parallelism configurations (TP/PP/DP/EP)
        and serving modes (aggregated and disaggregated) to find the deployment
        with the fewest GPUs that meets the constraints.

        Wraps `cli_recommend()` from the aisimulate SDK.

        Exactly one of `target_request_rate` or `target_concurrency` must be
        provided.

        Use the `include` query parameter to enrich each config with optional
        detail sections:
        - `config` — serving configuration (vLLM/SGLang parameters)
        - `memory` — per-component memory breakdown

        Example: `POST /recommend?include=config,memory`
      parameters:
        - name: include
          in: query
          required: false
          description: Comma-separated list of optional sections to include (config, memory).
          schema:
            type: string
            example: "config,memory"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendRequest'
      responses:
        '200':
          description: Recommendation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendResponse'
        '422':
          description: No feasible configuration found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /estimate:
    post:
      operationId: estimate
      tags: [sizing]
      summary: Single-point performance estimate for a given parallelism configuration
      description: |
        Given a model, GPU system, backend, and explicit parallelism settings
        (TP/PP/batch_size), returns predicted TTFT, TPOT, throughput, and memory.

        Use this when you know your deployment configuration and want to predict
        its performance. Use `/recommend` when you want to find the optimal config.

        Wraps `cli_estimate()` from the aisimulate SDK.

        Use the `include` query parameter to enrich the result with optional
        detail sections:
        - `config` — serving configuration (vLLM/SGLang parameters)
        - `memory` — per-component memory breakdown

        Example: `POST /estimate?include=config,memory`
      parameters:
        - name: include
          in: query
          required: false
          description: Comma-separated list of optional sections to include (config, memory).
          schema:
            type: string
            example: "config,memory"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateRequest'
      responses:
        '200':
          description: Performance estimate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        '422':
          description: Unsupported model/backend/system combination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RecommendRequest:
      type: object
      required: [model_path, system]
      properties:
        model_path:
          type: string
          description: HuggingFace model path or SDK model key.
          example: Qwen/Qwen3-32B-FP8
        system:
          type: string
          description: GPU system identifier.
          example: h200_sxm
        backend:
          type: string
          description: Inference backend.
          enum: [vllm, sglang, trtllm]
          default: vllm
        backend_version:
          type: string
          nullable: true
          description: Backend version. Null uses the latest available.
          default: null
        target_request_rate:
          type: number
          nullable: true
          description: Target request rate in req/s. Set exactly one of this or target_concurrency.
          default: null
        target_concurrency:
          type: number
          nullable: true
          description: Target concurrent users. Set exactly one of this or target_request_rate.
          default: null
        isl:
          type: integer
          description: Input sequence length (tokens).
          default: 4000
        osl:
          type: integer
          description: Output sequence length (tokens).
          default: 1000
        ttft:
          type: number
          description: Time to first token target (ms).
          default: 2000.0
        tpot:
          type: number
          description: Time per output token target (ms).
          default: 30.0
        request_latency:
          type: number
          nullable: true
          description: End-to-end request latency target (ms). Optional.
          default: null
        prefix:
          type: integer
          description: Prefix cache length (tokens).
          default: 0
        database_mode:
          type: string
          description: Performance database mode.
          enum: [HYBRID, SILICON, EMPIRICAL, SOL]
          default: HYBRID
        top_n:
          type: integer
          description: Number of top configurations to return.
          default: 5
          minimum: 1
          maximum: 20
        inclusive_tpot:
          type: boolean
          description: |
            Report TPOT as (ttft + tpot * (osl - 1)) / osl, spreading TTFT across all
            output tokens. Useful for comparing with benchmarks that report inclusive
            TPOT (e.g. GuideLLM).
          default: false
        model_config:
          type: object
          nullable: true
          description: Pre-fetched HuggingFace model config.json. Skips the HF download when supplied.
          default: null
    RecommendResponse:
      type: object
      required: [configs, chosen_mode]
      properties:
        configs:
          type: array
          description: Top configurations ranked by fewest GPUs, then throughput.
          items:
            $ref: '#/components/schemas/RecommendConfig'
        chosen_mode:
          type: string
          description: Serving mode chosen (agg or disagg).
          example: agg
    RecommendConfig:
      type: object
      properties:
        # GPU sizing
        total_gpus_needed:
          type: integer
          description: Total GPUs required (replicas x GPUs per worker).
        replicas_needed:
          type: integer
          description: Number of worker replicas.
        num_total_gpus:
          type: integer
          description: GPUs per worker (TP x PP x DP).

        # Parallelism
        tp:
          type: integer
          description: Tensor parallel size.
        pp:
          type: integer
          description: Pipeline parallel size.
        dp:
          type: integer
          description: Data parallel size.
        moe_tp:
          type: integer
          nullable: true
          description: MoE tensor parallel size (MoE models only).
        moe_ep:
          type: integer
          nullable: true
          description: MoE expert parallel size (MoE models only).
        cp:
          type: integer
          nullable: true
          description: Context parallel size.

        # Performance
        ttft:
          type: number
          description: Predicted TTFT (ms).
        tpot:
          type: number
          description: Predicted TPOT (ms).
        request_latency:
          type: number
          description: Predicted end-to-end request latency (ms).
        concurrency:
          type: integer
          description: Concurrency level at which these metrics hold.
        request_rate:
          type: number
          description: Achievable request rate (req/s).

        # Throughput
        tokens_per_second:
          type: number
          description: Total tokens/s.
        tokens_per_second_per_gpu:
          type: number
          description: Tokens/s per GPU.
        tokens_per_second_per_user:
          type: number
          description: Tokens/s per concurrent user.

        # Memory and metadata
        memory:
          type: number
          description: Memory usage (GB).
        model:
          type: string
          description: Resolved model identifier.
        system:
          type: string
          description: GPU system used.
        backend:
          type: string
          description: Backend used.
        backend_version:
          type: string
          description: Backend version used.

        # Quantization
        gemm:
          type: string
          description: GEMM quantization mode.
        kvcache:
          type: string
          description: KV cache quantization mode.
        fmha:
          type: string
          description: FMHA quantization mode.
        moe:
          type: string
          nullable: true
          description: MoE quantization mode.
        comm:
          type: string
          nullable: true
          description: Communication quantization mode.

        # Power
        power_w:
          type: number
          nullable: true
          description: Estimated power draw (watts).

        # Optional detail sections (present when requested via include= query param)
        serving_config:
          nullable: true
          description: Serving configuration (present when include=config).
          $ref: '#/components/schemas/ServingConfig'
        memory_breakdown:
          nullable: true
          description: Per-component memory breakdown (present when include=memory).
          $ref: '#/components/schemas/MemoryBreakdown'

        # Disaggregated detail (present for disagg results)
        prefill_config:
          nullable: true
          description: Prefill worker parallelism and serving config (disaggregated results only).
          $ref: '#/components/schemas/WorkerConfig'
        decode_config:
          nullable: true
          description: Decode worker parallelism and serving config (disaggregated results only).
          $ref: '#/components/schemas/WorkerConfig'
    WorkerConfig:
      type: object
      description: Parallelism and serving config for one worker role in a disaggregated deployment.
      properties:
        tp:
          type: integer
          nullable: true
          description: Tensor parallel size.
        pp:
          type: integer
          nullable: true
          description: Pipeline parallel size.
        dp:
          type: integer
          nullable: true
          description: Data parallel size.
        moe_tp:
          type: integer
          nullable: true
          description: MoE tensor parallel size.
        moe_ep:
          type: integer
          nullable: true
          description: MoE expert parallel size.
        num_workers:
          type: integer
          nullable: true
          description: Number of workers for this role.
        memory_gb:
          type: number
          nullable: true
          description: Memory usage (GB).
        gemm:
          type: string
          nullable: true
          description: GEMM quantization mode.
        kvcache:
          type: string
          nullable: true
          description: KV cache quantization mode.
        fmha:
          type: string
          nullable: true
          description: FMHA quantization mode.
        moe:
          type: string
          nullable: true
          description: MoE quantization mode.
        comm:
          type: string
          nullable: true
          description: Communication quantization mode.
        backend_version:
          type: string
          nullable: true
          description: Backend version used.
        memory_breakdown:
          nullable: true
          description: Per-component memory breakdown (present when include=memory).
          $ref: '#/components/schemas/MemoryBreakdown'
    ServingConfig:
      type: object
      description: Backend-specific serving parameters for the recommended configuration.
      properties:
        backend:
          type: string
          description: Inference backend.
        tensor_parallel_size:
          type: integer
        max_model_len:
          type: integer
          description: Maximum sequence length (ISL + OSL).
        max_num_seqs:
          type: integer
          description: Maximum concurrent sequences.
        gpu_memory_utilization:
          type: number
          description: Fraction of GPU memory to use (0.0-1.0).
        enable_chunked_prefill:
          type: boolean
        enable_prefix_caching:
          type: boolean
        quantization:
          type: string
          description: Quantization mode (fp8, int8, auto, etc).

    # ── Estimate ───────────────────────────────────────────────────────────
    EstimateRequest:
      type: object
      required: [model_path, system]
      properties:
        model_path:
          type: string
          description: HuggingFace model path or SDK model key.
          example: Qwen/Qwen3-32B
        system:
          type: string
          description: GPU system identifier.
          example: h200_sxm
        backend:
          type: string
          description: Inference backend.
          enum: [vllm, sglang, trtllm]
          default: vllm
        backend_version:
          type: string
          nullable: true
          description: Backend version. Null uses the latest available.
          default: null
        isl:
          type: integer
          description: Input sequence length (tokens).
          default: 4000
        osl:
          type: integer
          description: Output sequence length (tokens).
          default: 1000
        tp_size:
          type: integer
          description: Tensor parallel size.
          default: 1
        pp_size:
          type: integer
          description: Pipeline parallel size.
          default: 1
        batch_size:
          type: integer
          description: Batch size (max concurrent requests).
          default: 128
        database_mode:
          type: string
          description: Performance database mode.
          enum: [HYBRID, SILICON, EMPIRICAL, SOL]
          default: HYBRID
        gemm_quant_mode:
          type: string
          nullable: true
          default: null
        kvcache_quant_mode:
          type: string
          nullable: true
          default: null
        fmha_quant_mode:
          type: string
          nullable: true
          default: null
        moe_tp_size:
          type: integer
          nullable: true
          default: null
        moe_ep_size:
          type: integer
          nullable: true
          default: null
        attention_dp_size:
          type: integer
          default: 1
        inclusive_tpot:
          type: boolean
          description: |
            Report TPOT as (ttft + tpot * (osl - 1)) / osl, spreading TTFT across all
            output tokens. Useful for comparing with benchmarks that report inclusive
            TPOT (e.g. GuideLLM).
          default: false
        model_config:
          type: object
          nullable: true
          description: Pre-fetched HuggingFace model config.json. Skips the HF download when supplied.
          default: null
    EstimateResponse:
      type: object
      required: [ttft, tpot]
      description: Single-point performance estimate for a given parallelism configuration.
      properties:
        ttft:
          type: number
          description: Predicted TTFT (ms).
        tpot:
          type: number
          description: Predicted TPOT (ms).
        request_latency:
          type: number
          nullable: true
          description: Predicted end-to-end request latency (ms).
        tokens_per_second:
          type: number
          nullable: true
          description: Total tokens/s.
        tokens_per_second_per_gpu:
          type: number
          nullable: true
          description: Tokens/s per GPU.
        tokens_per_second_per_user:
          type: number
          nullable: true
          description: Tokens/s per concurrent user.
        memory:
          type: number
          nullable: true
          description: Memory usage (GB).
        concurrency:
          type: integer
          nullable: true
          description: Concurrency level (batch size) at which these metrics hold.
        tp:
          type: integer
          nullable: true
          description: Tensor parallel size.
        pp:
          type: integer
          nullable: true
          description: Pipeline parallel size.
        dp:
          type: integer
          nullable: true
          description: Data parallel size.
        system:
          type: string
          nullable: true
          description: GPU system used.
        backend:
          type: string
          nullable: true
          description: Backend used.
        backend_version:
          type: string
          nullable: true
          description: Backend version used.
        gemm:
          type: string
          nullable: true
          description: GEMM quantization mode.
        kvcache:
          type: string
          nullable: true
          description: KV cache quantization mode.
        power_w:
          type: number
          nullable: true
          description: Estimated power draw (watts).
        serving_config:
          nullable: true
          description: Serving configuration (present when include=config).
          $ref: '#/components/schemas/ServingConfig'
        memory_breakdown:
          nullable: true
          description: Per-component memory breakdown (present when include=memory).
          $ref: '#/components/schemas/MemoryBreakdown'

    # ── Memory ─────────────────────────────────────────────────────────────
...
    MemoryBreakdown:
      type: object
      properties:
        weights_bytes:
          type: integer
          description: Model weights memory (bytes).
        activations_bytes:
          type: integer
          description: Activation memory (bytes).
        runtime_overhead_bytes:
          type: integer
          description: Runtime overhead memory (bytes).
        comm_overhead_bytes:
          type: integer
          description: Communication buffer memory (bytes).
        kv_cache_bytes:
          type: integer
          description: KV cache memory (bytes). Mirrors total_kv_size_bytes.

    # ── Catalog ────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.